put utility methods in BlockWorkerUtils to NetworkUtils
